Add reconnect and every config attributes
We add them to the connection manager and the connection role, so we can pass
them to Redis when we connect. It could perhaps be a good idea to also allow
setting `conservative_reconnect`, but we are only dealing with messages here,
so I don't need it for my environment at least.